+++ homelab_git_management/run.sh
@@ -33,7 +33,7 @@
 
 set -e
 
-APP_VERSION="0.2.4"
+APP_VERSION="0.2.5"
 
 export APP_VERSION